revert moving
