+++ packages/wallet-sdk/src/sign/walletlink/WalletLinkSigner.test.ts
@@ -1,9 +1,9 @@
 import { fireEvent } from '@testing-library/preact';
 
-import { LOCAL_STORAGE_ADDRESSES_KEY } from './constants';
-import { MOCK_ADDERESS, MOCK_SIGNED_TX, MOCK_TX, MOCK_TYPED_DATA } from './mocks/fixtures';
-import { mockedWalletLinkRelay } from './mocks/relay';
-import { WalletLinkRelay } from './WalletLinkRelay';
+import { LOCAL_STORAGE_ADDRESSES_KEY } from './relay/constants';
+import { MOCK_ADDERESS, MOCK_SIGNED_TX, MOCK_TX, MOCK_TYPED_DATA } from './relay/mocks/fixtures';
+import { mockedWalletLinkRelay } from './relay/mocks/relay';
+import { WalletLinkRelay } from './relay/WalletLinkRelay';
 import { WalletLinkSigner } from './WalletLinkSigner';
 import { standardErrorCodes, standardErrors } from ':core/error';
 import { AddressString } from ':core/type';
@@ -20,9 +20,7 @@ const testStorage = new ScopedLocalStorage('walletlink', testWalletLinkUrl);
 
 const createAdapter = (options?: { relay?: WalletLinkRelay }) => {
   const adapter = new WalletLinkSigner({
-    appName: 'test',
-    appLogoUrl: null,
-    walletlinkUrl: testWalletLinkUrl,
+    metadata: { appName: 'test', appLogoUrl: null, appChainIds: [1] },
     updateListener: {
       onAccountsUpdate: () => {},
       onChainUpdate: () => {},
@@ -50,7 +48,7 @@ describe('LegacyProvider', () => {
     const spy = jest.spyOn(relay, 'resetAndReload');
 
     const provider = createAdapter({ relay });
-    await provider.close();
+    await provider.disconnect();
     expect(spy).toHaveBeenCalled();
   });
