+++ packages/wallet-sdk/src/sign/walletlink/relay/type/ServerMessage.ts
