+++ packages/wallet-sdk/src/sign/walletlink/relay/type/WalletLinkSessionConfig.ts
